+++ k256/src/arithmetic.rs
@@ -12,6 +12,7 @@ pub(crate) mod scalar;
 mod dev;
 
 pub use field::FieldElement;
+pub use mul::lincomb;
 
 use self::{affine::AffinePoint, projective::ProjectivePoint, scalar::Scalar};
 use crate::Secp256k1;