+++ p521/src/test_vectors.rs
@@ -1,0 +1,5 @@
+//! secp521r1 test vectors.
+
+#[cfg(test)]
+pub mod ecdsa;
+pub mod group;